Frontend/Tar/functions/requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
cryptography==43.0.0
deprecation==2.1.0
firebase-admin==6.5.0
firebase-functions==0.1.2
Flask==3.0.3
Flask-Cors==4.0.1
functions-framework==3.8.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.3
